Decouple auto-research worker seed from legacy core
