+++ loopx/capabilities/auto_research/legacy_core.py
@@ -9,13 +9,19 @@ are migrated or deleted.
 from __future__ import annotations
 
 import json
-import math
 import re
 import shlex
 from pathlib import Path
 from typing import Any, Iterable
 
 from . import evidence_packet as _evidence_packet
+from .quickstart_seed import (
+    AUTO_RESEARCH_DEFAULT_GOAL_ID,
+    AUTO_RESEARCH_DEFAULT_OBJECTIVE,
+    AUTO_RESEARCH_QUICKSTART_SCHEMA_VERSION,
+    AUTO_RESEARCH_QUICKSTART_TEMPLATE,
+    build_auto_research_quickstart,
+)
 from ...visible_multi_agent_launcher import (
     build_visible_lane_command,
     build_visible_multi_agent_payload,
@@ -51,15 +57,11 @@ RESEARCH_SHOWCASE_PROJECTION_SCHEMA_VERSION = "research_showcase_projection_v0"
 AUTO_RESEARCH_PROJECTION_SCHEMA_VERSION = "decentralized_auto_research_projection_v0"
 AUTO_RESEARCH_ARTIFACT_PACKET_SCHEMA_VERSION = "auto_research_artifact_packet_v0"
 AUTO_RESEARCH_BOARD_SCHEMA_VERSION = "auto_research_frontstage_board_v0"
-AUTO_RESEARCH_QUICKSTART_SCHEMA_VERSION = "auto_research_quickstart_v0"
 AUTO_RESEARCH_DEMO_SUPERVISOR_SCHEMA_VERSION = "auto_research_demo_supervisor_plan_v0"
 AUTO_RESEARCH_DEMO_ACCEPTANCE_PACKET_SCHEMA_VERSION = "auto_research_demo_acceptance_packet_v0"
 AUTO_RESEARCH_DEMO_E2E_SCHEMA_VERSION = "auto_research_demo_e2e_result_v0"
 AUTO_RESEARCH_ROLE_PROFILE_SCHEMA_VERSION = "auto_research_role_profile_v0"
 AUTO_RESEARCH_PUBLIC_CLAIM_BOUNDARY_SCHEMA_VERSION = "auto_research_public_claim_boundary_v0"
-AUTO_RESEARCH_DEFAULT_GOAL_ID = "loopx-auto-research-knn"
-AUTO_RESEARCH_DEFAULT_OBJECTIVE = "Improve exact k-nearest-neighbor inference under a protected evaluator."
-AUTO_RESEARCH_QUICKSTART_TEMPLATE = "knn-exact"
 ROLLOUT_EVIDENCE_GRAPH_SOURCE_KIND = "loopx_rollout_event_log"
 AUTO_RESEARCH_DEMO_DEFAULT_LANES = (
     (
@@ -377,20 +379,6 @@ def validate_auto_research_fixture(payload: dict[str, Any]) -> dict[str, Any]:
         "raw_logs_recorded": False,
         "private_artifacts_recorded": False,
     }
-
-
-def _relative_pack_dir(value: Any) -> Path:
-    text = _compact_public_text(value, field="output_dir", max_len=160)
-    if "\\" in text:
-        raise ValueError("output_dir must use forward-slash relative paths")
-    path = Path(text)
-    if path.is_absolute():
-        raise ValueError("output_dir must be relative")
-    if not path.parts:
-        raise ValueError("output_dir must be non-empty")
-    if any(part in {"", ".", ".."} for part in path.parts):
-        raise ValueError("output_dir must not contain current or parent-directory markers")
-    return path
 
 
 def _shell_arg(value: str) -> str:
@@ -1301,418 +1289,6 @@ def build_auto_research_demo_acceptance_packet(
         },
     }
 
-
-def _quickstart_contract(*, goal_id: str, objective: str) -> dict[str, Any]:
-    contract = _validate_research_contract(
-        {
-            "schema_version": _EP_RESEARCH_CONTRACT_SCHEMA_VERSION,
-            "goal_id": goal_id,
-            "research_objective": objective,
-            "editable_scope": ["solution_candidate.py"],
-            "protected_scope": [
-                "protected_eval.py",
-                "solution_baseline.py",
-                "research_contract.json",
-            ],
-            "metric": {
-                "name": "deterministic_speedup",
-                "direction": "maximize",
-                "baseline": 1.0,
-            },
-            "dev_eval": "python3 protected_eval.py --solution solution_candidate.py --split dev",
-            "holdout_eval": "python3 protected_eval.py --solution solution_candidate.py --split holdout",
-            "promotion_policy": "requires_dev_and_holdout_improvement_exactness_and_clean_boundary",
-        }
-    )
-    contract["no_upload"] = True
-    return contract
-
-
-_QUICKSTART_PROTECTED_EVAL = '''#!/usr/bin/env python3
-from __future__ import annotations
-
-import argparse
-import importlib.util
-import json
-import math
-import sys
-from pathlib import Path
-from typing import Any
-
-
-sys.dont_write_bytecode = True
-
-SCHEMA_VERSION = "auto_research_knn_eval_result_v0"
-PACK_DIR = Path(__file__).resolve().parent
-Point = tuple[float, ...]
-
-SPLITS: dict[str, dict[str, int]] = {
-    "dev": {"seed": 17, "train_count": 256, "query_count": 18, "dims": 4, "k": 3},
-    "holdout": {"seed": 31, "train_count": 512, "query_count": 24, "dims": 4, "k": 3},
-}
-
-
-def _point(seed: int, index: int, dims: int) -> Point:
-    values = []
-    for dim in range(dims):
-        raw = (seed * (dim + 5) + index * (dim * 23 + 11) + index * index * (dim + 3)) % 997
-        values.append(raw / 97.0)
-    return tuple(values)
-
-
-def build_split(split: str) -> tuple[list[Point], list[Point], int, dict[str, int]]:
-    if split not in SPLITS:
-        raise ValueError(f"unknown split {split!r}")
-    spec = SPLITS[split]
-    train = [_point(spec["seed"], index, spec["dims"]) for index in range(spec["train_count"])]
-    queries = [
-        _point(spec["seed"] + 101, index, spec["dims"])
-        for index in range(spec["query_count"])
-    ]
-    return train, queries, spec["k"], spec
-
-
-def _squared_distance(left: Point, right: Point) -> float:
-    return sum((a - b) * (a - b) for a, b in zip(left, right))
-
-
-def oracle_knn(train: list[Point], queries: list[Point], k: int) -> list[list[int]]:
-    expected: list[list[int]] = []
-    for query in queries:
-        ranked = sorted((_squared_distance(query, point), index) for index, point in enumerate(train))
-        expected.append([index for _, index in ranked[:k]])
-    return expected
-
-
-def load_solution(path: Path) -> Any:
-    spec = importlib.util.spec_from_file_location("auto_research_knn_solution", path)
-    if spec is None or spec.loader is None:
-        raise ValueError(f"cannot load solution from {path}")
-    module = importlib.util.module_from_spec(spec)
-    spec.loader.exec_module(module)
-    if not hasattr(module, "solve_knn"):
-        raise ValueError("solution must define solve_knn(train, queries, k)")
-    return module
-
-
-def ranking_work_units(strategy: str, *, train_count: int, query_count: int, k: int) -> int:
-    if strategy == "partial_selection":
-        rank_factor = max(1, math.ceil(math.log2(k + 1)))
-    else:
-        rank_factor = max(1, math.ceil(math.log2(train_count)))
-    return query_count * train_count * rank_factor
-
-
-def evaluate(solution_path: Path, split: str) -> dict[str, Any]:
-    solution_path = solution_path.resolve()
-    train, queries, k, spec = build_split(split)
-    module = load_solution(solution_path)
-    strategy = str(getattr(module, "STRATEGY", "unknown"))
-    expected = oracle_knn(train, queries, k)
-    actual = module.solve_knn(train, queries, k)
-    exact = actual == expected
-
-    baseline_units = ranking_work_units(
-        "full_sort",
-        train_count=spec["train_count"],
-        query_count=spec["query_count"],
-        k=k,
-    )
-    candidate_units = ranking_work_units(
-        strategy,
-        train_count=spec["train_count"],
-        query_count=spec["query_count"],
-        k=k,
-    )
-    speedup = baseline_units / candidate_units if exact else None
-    improved = bool(speedup is not None and speedup > 1.0)
-    protected_scope_clean = solution_path.parent == PACK_DIR and solution_path.name in {
-        "solution_baseline.py",
-        "solution_candidate.py",
-    }
-    promotion_ready = exact and improved and protected_scope_clean
-    return {
-        "schema_version": SCHEMA_VERSION,
-        "split": split,
-        "solution": solution_path.name,
-        "strategy": strategy,
-        "dataset": {
-            "train_count": spec["train_count"],
-            "query_count": spec["query_count"],
-            "dims": spec["dims"],
-            "k": k,
-            "seed": spec["seed"],
-        },
-        "metric": {
-            "name": "deterministic_speedup",
-            "direction": "maximize",
-            "value": round(speedup, 6) if speedup is not None else None,
-            "baseline": 1.0,
-        },
-        "work_units": {
-            "baseline_full_sort": baseline_units,
-            "candidate": candidate_units,
-        },
-        "exact": exact,
-        "protected_scope_clean": protected_scope_clean,
-        "no_upload": True,
-        "eval_status": "scored" if exact else "guardrail_failed",
-        "primary_metric_status": "improved" if improved else ("baseline" if exact else "failed"),
-        "promotion_gate": {
-            "requires": [
-                "exact_neighbor_identity",
-                "dev_and_holdout_improvement",
-                "protected_scope_clean",
-                "no_upload",
-            ],
-            "ready_for_split": promotion_ready,
-        },
-        "artifact_refs": [
-            f"knn_pack:{split}:{strategy}",
-        ],
-    }
-
-
-def main() -> int:
-    parser = argparse.ArgumentParser(description="Protected evaluator for the public LoopX auto-research k-NN pack.")
-    parser.add_argument("--solution", required=True, help="Path to a solution module.")
-    parser.add_argument("--split", choices=sorted(SPLITS), required=True)
-    args = parser.parse_args()
-    payload = evaluate(Path(args.solution), args.split)
-    print(json.dumps(payload, indent=2, sort_keys=True))
-    return 0 if payload["exact"] and payload["protected_scope_clean"] else 1
-
-
-if __name__ == "__main__":
-    raise SystemExit(main())
-'''
-
-
-_QUICKSTART_SOLUTION_BASELINE = '''from __future__ import annotations
-
-
-STRATEGY = "full_sort"
-
-
-Point = tuple[float, ...]
-
-
-def _squared_distance(left: Point, right: Point) -> float:
-    return sum((a - b) * (a - b) for a, b in zip(left, right))
-
-
-def solve_knn(train: list[Point], queries: list[Point], k: int) -> list[list[int]]:
-    """Reference exact k-NN solver using a full distance sort per query."""
-
-    results: list[list[int]] = []
-    for query in queries:
-        ranked = sorted((_squared_distance(query, point), index) for index, point in enumerate(train))
-        results.append([index for _, index in ranked[:k]])
-    return results
-'''
-
-
-_QUICKSTART_SOLUTION_CANDIDATE = '''from __future__ import annotations
-
-import heapq
-
-
-STRATEGY = "partial_selection"
-
-
-Point = tuple[float, ...]
-
-
-def _squared_distance(left: Point, right: Point) -> float:
-    return sum((a - b) * (a - b) for a, b in zip(left, right))
-
-
-def solve_knn(train: list[Point], queries: list[Point], k: int) -> list[list[int]]:
-    """Exact k-NN using partial selection instead of sorting every distance."""
-
-    results: list[list[int]] = []
-    for query in queries:
-        nearest = heapq.nsmallest(
-            k,
-            ((_squared_distance(query, point), index) for index, point in enumerate(train)),
-        )
-        results.append([index for _, index in nearest])
-    return results
-'''
-
-
-_QUICKSTART_README = '''# LoopX Auto Research k-NN Pack
-
-This generated pack is a public-safe starter for a decentralized auto-research
-run. Edit only `solution_candidate.py`; keep `protected_eval.py`,
-`solution_baseline.py`, and `research_contract.json` unchanged.
-
-Run dev:
-
-```bash
-python3 protected_eval.py --solution solution_candidate.py --split dev
-```
-
-Run holdout:
-
-```bash
-python3 protected_eval.py --solution solution_candidate.py --split holdout
-```
-
-Promotion requires exact neighbor identity, dev and holdout improvement, clean
-protected scope, and no upload/private artifacts.
-'''
-
-
-def _quickstart_template_files(contract: dict[str, Any]) -> dict[str, str]:
-    return {
-        "research_contract.json": json.dumps(contract, indent=2, sort_keys=True) + "\n",
-        "protected_eval.py": _QUICKSTART_PROTECTED_EVAL,
-        "solution_baseline.py": _QUICKSTART_SOLUTION_BASELINE,
-        "solution_candidate.py": _QUICKSTART_SOLUTION_CANDIDATE,
-        "README.md": _QUICKSTART_README,
-    }
-
-
-def _quickstart_file_summary(
-    *,
-    pack_dir: Path,
-    files: dict[str, str],
-    write_status: str,
-) -> list[dict[str, Any]]:
-    protected_names = {"protected_eval.py", "solution_baseline.py", "research_contract.json"}
-    role_by_name = {
-        "README.md": "operator_notes",
-        "protected_eval.py": "protected_evaluator",
-        "research_contract.json": "research_contract",
-        "solution_baseline.py": "protected_baseline",
-        "solution_candidate.py": "editable_candidate",
-    }
-    return [
-        {
-            "path": f"{pack_dir.as_posix()}/{name}",
-            "role": role_by_name.get(name, "pack_file"),
-            "protected": name in protected_names,
-            "write_status": write_status,
-        }
-        for name in sorted(files)
-    ]
-
-
-def build_auto_research_quickstart(
-    *,
-    agent_id: str,
-    goal_id: str = AUTO_RESEARCH_DEFAULT_GOAL_ID,
-    objective: str = AUTO_RESEARCH_DEFAULT_OBJECTIVE,
-    output_dir: str = "auto_research_knn_pack",
-    template: str = AUTO_RESEARCH_QUICKSTART_TEMPLATE,
-    execute: bool = False,
-    cwd: Path | None = None,
-) -> dict[str, Any]:
-    """Create or preview the public k-NN auto-research starter contract.
-
-    The default mode is read-only. `execute=True` writes a protected evaluator
-    pack below the current working directory and refuses to overwrite an
-    existing pack, so heartbeat agents can preview the next hypothesis without
-    producing local artifacts by accident.
-    """
-
-    agent = _compact_public_token(agent_id, field="agent_id")
-    goal = _compact_public_token(goal_id, field="goal_id")
-    objective_text = _compact_public_text(objective, field="objective", max_len=240)
-    selected_template = _compact_public_token(template, field="template")
-    if selected_template != AUTO_RESEARCH_QUICKSTART_TEMPLATE:
-        raise ValueError(f"template must be {AUTO_RESEARCH_QUICKSTART_TEMPLATE}")
-    pack_dir = _relative_pack_dir(output_dir)
-    contract = _quickstart_contract(goal_id=goal, objective=objective_text)
-    files = _quickstart_template_files(contract)
-    dev_command = (
-        f"python3 {pack_dir.as_posix()}/protected_eval.py "
-        f"--solution {pack_dir.as_posix()}/solution_candidate.py --split dev"
-    )
-    holdout_command = (
-        f"python3 {pack_dir.as_posix()}/protected_eval.py "
-        f"--solution {pack_dir.as_posix()}/solution_candidate.py --split holdout"
-    )
-    hypothesis = _validate_research_hypothesis(
-        {
-            "schema_version": _EP_RESEARCH_HYPOTHESIS_SCHEMA_VERSION,
-            "hypothesis_id": "hyp_quickstart_partial_selection",
-            "todo_id": "todo_auto_research_quickstart_001",
-            "claimed_by": agent,
-            "mechanism_family": "partial_selection",
-            "hypothesis": "Use exact partial selection to avoid full distance sorting.",
-            "status": "active",
-            "grounding_refs": ["quickstart:knn_exact_pack"],
-            "blocked_by": [],
-        }
-    )
-    write_status = "would_write"
-    if execute:
-        root = (cwd or Path.cwd()).resolve()
-        target_dir = (root / pack_dir).resolve()
-        try:
-            target_dir.relative_to(root)
-        except ValueError as exc:
-            raise ValueError("output_dir must resolve inside the current working directory") from exc
-        if target_dir.exists():
-            raise ValueError(f"output_dir already exists: {pack_dir.as_posix()}")
-        target_dir.mkdir(parents=True)
-        for name, contents in files.items():
-            target = target_dir / name
-            target.write_text(contents, encoding="utf-8")
-            if name == "protected_eval.py":
-                target.chmod(0o755)
-        write_status = "created"
-    file_summary = _quickstart_file_summary(
-        pack_dir=pack_dir,
-        files=files,
-        write_status=write_status,
-    )
-    return {
-        "ok": True,
-        "schema_version": AUTO_RESEARCH_QUICKSTART_SCHEMA_VERSION,
-        "mode": "execute" if execute else "dry_run",
-        "template": selected_template,
-        "research_contract": contract,
-        "pack_dir": pack_dir.as_posix(),
-        "files": file_summary,
-        "next_runnable_hypothesis": hypothesis | {
-            "allowed_action": "run_dev_attempt",
-            "run_command": dev_command,
-        },
-        "next_commands": [
-            {
-                "label": "create_pack",
-                "command": (
-                    f"loopx --format json auto-research quickstart --agent-id {agent} "
-                    f"--goal-id {goal} --output-dir {pack_dir.as_posix()} --execute"
-                ),
-                "required_when": "dry_run",
-            },
-            {"label": "run_dev", "command": dev_command},
-            {"label": "run_holdout", "command": holdout_command},
-            {
-                "label": "record_evidence",
-                "command": (
-                    "loopx --format json auto-research evidence "
-                    f"--contract {pack_dir.as_posix()}/research_contract.json "
-                    "--eval-result dev-result.public.json --eval-result holdout-result.public.json "
-                    "--hypothesis-id hyp_quickstart_partial_selection "
-                    "--todo-id todo_auto_research_quickstart_001 "
-                    f"--agent-id {agent} --claimed-by {agent} "
-                    "--mechanism-family partial_selection "
-                    '--hypothesis "Use exact partial selection to avoid full distance sorting."'
-                ),
-            },
-        ],
-        "public_boundary": {
-            "raw_logs_recorded": False,
-            "private_artifacts_recorded": False,
-            "absolute_paths_recorded": False,
-            "source": "generated_public_quickstart_contract",
-        },
-    }
 
 
 def build_auto_research_projection(
